refactor folders name
