+++ src/common/components/index.ts
@@ -2,4 +2,4 @@ export { MarkDownViewerComponent } from './markdownViewer';
 export { PanelComponent, PanelItem } from './panel';
 export { Breadcrumb, BreadcrumbItem } from './breadcrumb';
 export { CounterButton } from './counterButton/counterButton';
-export { SubscriptionManager } from './serverRequestManager/serverRequestManager';
+export { SubscriptionManager } from './subscriptionManager/subscriptionManager';
